cmd: ctr: metrics: factor out cgroup1/cgroup2/wstats specifics
Factoring out specifics for cgroup1 vs cgroup2 vs wstats specicial handling
into separate source files, so theylater can be made build-time optional.
